Refactor the 'convertToRender' command
