+++ astraledit-bukkit-plugin/src/main/java/com/github/shynixn/astraledit/bukkit/logic/business/SelectionCommandExecutor.java
@@ -42,6 +42,7 @@ public class SelectionCommandExecutor extends SimpleCommandExecutor.Registered {
         this.commands.add(new MirrorCommand(this.plugin, manager));
         this.commands.add(new PlaceCommand(this.plugin, manager));
         this.commands.add(new FlipCommand(this.plugin, manager));
+        this.commands.add(new ConvertToRenderCommand(this.plugin));
     }
 
     /**
@@ -125,26 +126,6 @@ public class SelectionCommandExecutor extends SimpleCommandExecutor.Registered {
     }
 
     //ASYNC
-
-    /**
-     * Converts a rendered object to blocks
-     *
-     * @param player player
-     */
-    private void convertToRenderCommand(final Player player) {
-        this.runAsyncTask(() -> {
-            try {
-                player.sendMessage(AstralEditPlugin.PREFIX_SUCCESS + "Removing blocks and rendering selection asynchronously...");
-                this.plugin.getServer().getScheduler().runTaskAsynchronously(this.plugin, () -> {
-                    AstralEditApi.INSTANCE.renderAndDestroy(player);
-                    player.sendMessage(AstralEditPlugin.PREFIX_SUCCESS + "Finished converting selection.");
-                });
-            } catch (final Exception e) {
-                player.sendMessage(AstralEditPlugin.PREFIX_ERROR + "Failed converting WE selection!");
-                player.sendMessage(AstralEditPlugin.PREFIX_ERROR + "Check if you selected an area with Worldedit.");
-            }
-        });
-    }
 
     /**
      * Converts the blocks to a rendered object
